Cross platform command line (#13) +semver: major
We've ported the existing command line tool to .NET Core, so that it can work cross-platform on Windows, Mac and Linux. This is the first big step in a longer story to make Augurk more accessible for teams that typically do not work on Windows.

As part of this, we've also rethought the distribution of the command line. We no longer ship it as a NuGet package containing a bunch of binaries. Instead we've made it available as a so called .NET Core Global Tool on NuGet, as well as platform specific zips that can be used in CI/CD pipelines which do not require .NET Core to be installed.

Fixes #13
Fixes #12
Fixes #11